% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhjboot.R
\name{hhjboot}
\alias{hhjboot}
\title{HHJ Algorithm}
\usage{
hhjboot(
  series,
  nb = 100L,
  n.iter = 10L,
  pilot.block.length = NULL,
  sub.block.size = NULL,
  bofb = 1,
  search.grid = NULL,
  grid.step = 1,
  cl = NULL
)
}
\arguments{
\item{series}{a numeric vector or time series giving the original data for which to find the optimal block length for.}

\item{nb}{number of bootstrap series to compute.}

\item{n.iter}{number of iterations for HHJ algorithm.}

\item{pilot.block.length}{pilot block length (l\* __in HHJ__) for which to perform initial block bootstraps.}

\item{sub.block.size}{length of each overlapping subsample (m __in HHJ__).}

\item{bofb}{length of the basic blocks in the __block of blocks__ bootstrap.}

\item{search.grid}{the range of solutions around l* to evaluate within the MSE function after 1st iteration.}

\item{grid.step}{if grid.step = 1 then each block length will be evaluated in the MSE function, if grid.step > 1, the the MSE function will search over all the sequence of block lengths from 1 to m by grid.step.}

\item{cl}{a cluster object, created by this package parallel or by package snow. If NULL, use the registered default cluster.}
}
\description{
Perform the HHJ algorithm to select the optimal block length (l) for the
moving block bootstrap.
}
\examples{

hhjboot(stats::arima.sim(list(order = c(1, 0, 0), ar = 0.5),
  n = 500, innov = rnorm(500)
))
}
