% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hhjboot.R
\name{plot.hhjboot}
\alias{plot.hhjboot}
\title{Plot MSE Function for HHJ Algorithm}
\usage{
\method{plot}{hhjboot}(x, iter = NULL, ...)
}
\arguments{
\item{x}{an object of class "hhjboot"}

\item{iter}{a vector of \code{hhjboot()} iterations to plot. \pkg{NULL}, by
default will plot all iterations}

\item{...}{Arguments to be passed to methods}
}
\description{
Plot MSE Function for HHJ Algorithm
}
\examples{
\dontrun{
# Generate AR(1) time series
sim <- stats::arima.sim(list(order = c(1, 0, 0), ar = 0.5),
                        n = 500, innov = rnorm(500))

# Generate 'hhjboot' class object of optimal block length for series
hhj <- hhjboot(sim, sub.block.size = 10)

## S3 method for class 'hhjboot'
plot(hhj)
}

}
