% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwsd.R
\name{pwsd}
\alias{pwsd}
\title{Politis and White (2004) Spectral Density Automatic Block-Length Selection}
\usage{
pwsd(
  data,
  K_N = NULL,
  M_max = NULL,
  b_max = NULL,
  c = NULL,
  round = FALSE,
  correlogram = TRUE
)
}
\arguments{
\item{data}{an \eqn{n x k} data.frame or matrix where the optimal block-length
will be computed for each of the \eqn{k} columns.}

\item{K_N}{the maximum lags for which to apply the \emph{implied hypothesis} testing
of the autocorrelations \eqn{rho_k}. See Politis and White (2004) <INSERT>}

\item{M_max}{the upper bound for the optimal number of lags \eqn{M} to
compute the auto-covariance. \emph{See} Theorem 3.3 (ii) of Politis and White}

\item{b_max}{the upper bound for the optimal block-length. Defaults to
\code{ceiling(min(3 * sqrt(n), n / 3))} per Politis and White (2004).}

\item{c}{constant which acts as the significance level for implied
hypothesis test. Defaults to \code{qnorm(0.975)} for a 95% confidence level.
Politis and  White (2004) suggest \code{c=2}.}

\item{round}{logical. Set to \code{FALSE} by default. Setting to \code{TRUE}
will round the final estimates for the optimal block-length.}

\item{correlogram}{logical. Setting \code{TRUE} will output a plot
of the correlogram (\emph{i.e.} a plot of \eqn{R(k)} vs. \eqn{k}) to the console.}
}
\value{
an object of class 'pwsd'
}
\description{
Run the Automatic Block-Length selection method proposed by Politis and White
 (2004) and corrected in Patton, Politis, and White (2009). The method is
 based on spectral density estimate via flat-top lag windows of Politis and
 Romano (1995).
}
