% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hhjboot.R
\name{plot.hhjboot}
\alias{plot.hhjboot}
\title{Plot MSE Function for HHJ Algorithm}
\usage{
\method{plot}{hhjboot}(x, iter = NULL, ...)
}
\arguments{
\item{x}{an object of class 'hhjboot'}

\item{iter}{a vector of \code{hhjboot()} iterations to plot. \code{NULL}. All
iterations are plotted by default.}

\item{...}{
  Arguments passed on to \code{\link[base:plot]{base::plot}}
  \describe{
    \item{\code{y}}{the y coordinates of points in the plot, \emph{optional}
    if \code{x} is an appropriate structure.}
  }}
}
\description{
S3 Method for objects of class 'hhjboot'
}
\examples{

\dontrun{
# Generate AR(1) time series
sim <- stats::arima.sim(list(order = c(1, 0, 0), ar = 0.5),
                        n = 500, innov = rnorm(500))

# Generate 'hhjboot' class object of optimal block length for series
hhj <- hhjboot(sim, sub.block.size = 10)

## S3 method for class 'hhjboot'
plot(hhj)
}

}
