% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hhjboot.R
\name{plot.hhjboot}
\alias{plot.hhjboot}
\title{Plot MSE Function for HHJ Algorithm}
\usage{
\method{plot}{hhjboot}(x, ..., iter = NULL)
}
\arguments{
\item{x}{an object of class "hhjboot"}

\item{...}{Arguments to be passed to methods}

\item{iter}{a vector of \code{hhjboot()} iterations to plot. \pkg{NULL}, by
default will plot all iterations}
}
\description{
Plot MSE Function for HHJ Algorithm
}
